Renamed files to camel case
Some files edited to allow for changed name of CSS files or dependencies
