+++ src/components/SlideControl/slideControl.module.css
